refactor(parser): extract lowering phase from oracle_effect/mod.rs into lower.rs (#1648)
Enforces the documented parse/lower module boundary (D-08). The lowering
phase — `lower_effect_chain_ir` and its 92 helpers (~4,411 lines) — is
now in its own file, `oracle_effect/lower.rs`. The module is a pure
`&EffectChainIr → AbilityDefinition` transformation with no `ParseContext`
dependency, making its contract explicit and separately testable.

Rebased onto upstream/main (d780d8b0) which added 17 new functions to
the lower section since the original extraction. All are included.

`mod.rs` shrinks from 43,876 to 39,489 lines. All public API
(`lower_effect_chain_ir`, `capitalize`, `strip_trailing_duration`,
`strip_trailing_where_x`, `parse_with_counters_suffix`,
`parse_counter_suffix_body_combinator`, `parse_controls_permanent_object`)
is re-exported from `mod.rs` so no external call-site changes are needed.

All 10,013 engine tests pass.
